requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
Django==4.1.3
djangorestframework==3.14.0
mysqlclient==2.1.1
pycparser==2.21
pyOpenSSL==22.1.0
pytz==2022.6
sqlparse==0.4.3
